修改JS事件拦截冲突 int --> float

+++ common_book/src/main/java/com/intelligent/reader/activity/LabelsDetailActivity.java
@@ -534,7 +534,7 @@ public class LabelsDetailActivity extends FrameActivity implements View.OnClickL
         if (isNeedInterceptSlide()) {
             jsInterfaceHelper.setOnH5PagerInfo(new JSInterfaceHelper.OnH5PagerInfoListener() {
                 @Override
-                public void onH5PagerInfo(int x, int y, int width, int height) {
+                public void onH5PagerInfo(float x, float y, float width, float height) {
                     mPagerDesc = new PagerDesc(y, x, x + width, y + height);
                 }
             });
@@ -706,8 +706,8 @@ public class LabelsDetailActivity extends FrameActivity implements View.OnClickL
                                 h5Margin = loction[1];
                             }
                             if (null != mPagerDesc) {
-                                int top = mPagerDesc.getTop();
-                                int bottom = top + (mPagerDesc.getBottom() - mPagerDesc.getTop());
+                                float top = mPagerDesc.getTop();
+                                float bottom = top + (mPagerDesc.getBottom() - mPagerDesc.getTop());
                                 DisplayMetrics metric = getResources().getDisplayMetrics();
                                 top = (int) (top * metric.density) + h5Margin;
                                 bottom = (int) (bottom * metric.density) + h5Margin;
